feat(match): implement match user check and display logic
